refactor logs add warning handler
